filter_aws: ec2 tags include/exclude option
Add configuration for including EC2 tags.

Previously fluent-bit included all found EC2 tags, which in
certain scenarios might not be acceptable (e.g. unnecessarily bloating
space in case tags are 'heavy').

This commit allows to configure which specific tags should be included,
with two possible exclusive configuration options:
 - tags_include
 - tags_exclude

tags_include and tags_exclude are configuration options for aws filter
plugin. Value type is a string, which should be represented as tag keys
joined with a ',' character. For example: 'tag1,tag2'.

tags_include allows to define a list of specific tags which, if found in
the metadata server, will be injected into the logs. All tags which are
not in the list will be ignored (won't be injected to the log).

tags_exclude allows to define a list of tags to be excluded. These tags
will not be included in the log. tags_exclude has any effect only if
tags_include is not defined.

Signed-off-by: Mateusz Warzyński <mateusz@warzynski.pro>